egg_sketches
Module recursive_extract

Alternative recursive descent implementation of sketch extraction inspired from pattern matching. This is instead of the bottom-up e-class analysis approach.

FIXME: recursive descent gives wrong results since update to egg main!

Functions

eclass_extract_sketch

Returns the best program satisfying s according to cost_f that is represented in the id e-class of egraph, if it exists.